Finished linting changes to /frontend
