+++ frontend/src/utils/FirebaseConfig.tsx
@@ -1,0 +1,15 @@
+/**
+ * Config information for Firebase.
+ * May be moved to environmental variables later.
+ */
+export const firebaseConfig = {
+  apiKey: "AIzaSyBssbaMlxIJHYI7G7zOriU0VaWGnGrQv5M",
+  authDomain: "low-price-center.firebaseapp.com",
+  projectId: "low-price-center",
+  storageBucket: "low-price-center.firebasestorage.app",
+  messagingSenderId: "163704233704",
+  appId: "1:163704233704:web:6ee0dc540f6f25d6ceb35d",
+  measurementId: "G-RV7RV9W17W",
+};
+
+export default firebaseConfig;
