+++ frontend/src/utils/FirebaseProvider.tsx
@@ -3,6 +3,8 @@ import { GoogleAuthProvider, User, getAuth, signInWithPopup, signOut } from "fir
 import { MouseEventHandler, ReactNode, createContext, useEffect, useState } from "react";
 import { get, post } from "src/api/requests";
 
+import { firebaseConfig } from "src/utils/FirebaseConfig";
+
 /**
  * Context used by FirebaseProvider to provide app and user to pages
  */
@@ -19,20 +21,6 @@ const FirebaseContext = createContext<{
   openGoogleAuthentication: () => {},
   signOutFromFirebase: () => {},
 });
-
-/**
- * Config information for Firebase.
- * May be moved to environmental variables later.
- */
-export const firebaseConfig = {
-  apiKey: "AIzaSyBssbaMlxIJHYI7G7zOriU0VaWGnGrQv5M",
-  authDomain: "low-price-center.firebaseapp.com",
-  projectId: "low-price-center",
-  storageBucket: "low-price-center.firebasestorage.app",
-  messagingSenderId: "163704233704",
-  appId: "1:163704233704:web:6ee0dc540f6f25d6ceb35d",
-  measurementId: "G-RV7RV9W17W",
-};
 
 /**
  * Wraps children in FirebaseContext.Provider to give all
